Moved JSONResponse class to io.rest bundle fixes #993
