+++ bundles/automation/org.eclipse.smarthome.automation.rest/META-INF/MANIFEST.MF
@@ -15,7 +15,6 @@ Import-Package: io.swagger.annotations;resolution:=optional,
  org.eclipse.smarthome.core.common,
  org.eclipse.smarthome.core.common.registry,
  org.eclipse.smarthome.io.rest,
- org.eclipse.smarthome.io.rest.core,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
