+++ bundles/io/org.eclipse.smarthome.io.rest/META-INF/MANIFEST.MF
@@ -7,13 +7,14 @@ Bundle-Vendor: Eclipse.org/SmartHome
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Import-Package: com.google.common.base,
  com.google.common.collect,
+ com.google.gson,
  io.swagger.annotations;resolution:=optional,
  javax.servlet,
  javax.servlet.http,
  javax.ws.rs,
  javax.ws.rs.container,
  javax.ws.rs.core,
- javax.ws.rs.ext;version="1.1.1",
+ javax.ws.rs.ext,
  org.apache.commons.lang,
  org.eclipse.smarthome.core.events,
  org.eclipse.smarthome.core.items,
