+++ bundles/io/org.eclipse.smarthome.io.rest/src/main/java/org/eclipse/smarthome/io/rest/JSONResponse.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.io.rest.core;
+package org.eclipse.smarthome.io.rest;
 
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
